+++ ezdxf/dxf2html.py
@@ -11,7 +11,7 @@ import os
 import io
 
 from ezdxf import readfile
-from ezdxf.tags import tag_type, point_tuple, is_point_code
+from ezdxf.dxftag import tag_type, point_tuple, is_point_code
 from ezdxf.c23 import escape, ustr
 from ezdxf.reflinks import get_reference_link
 from ezdxf.sections import KNOWN_SECTIONS